+++ install/empty_data.php
@@ -5617,11 +5617,11 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_SELF_SERVICE,
                 'name' => 'followup',
-                'rights' => READ | CREATE,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::ADDMY,
             ], [
                 'profiles_id' => self::PROFILE_SELF_SERVICE,
                 'name' => 'task',
-                'rights' => READ,
+                'rights' => CommonITILTask::SEEPUBLIC,
             ], [
                 'profiles_id' => self::PROFILE_SELF_SERVICE,
                 'name' => 'planning',
@@ -5893,11 +5893,11 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_OBSERVER,
                 'name' => 'followup',
-                'rights' => READ | CREATE,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::ADDMY,
             ], [
                 'profiles_id' => self::PROFILE_OBSERVER,
                 'name' => 'task',
-                'rights' => READ,
+                'rights' => CommonITILTask::SEEPUBLIC,
             ], [
                 'profiles_id' => self::PROFILE_TECHNICIAN,
                 'name' => 'projecttask',
@@ -6190,13 +6190,13 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_ADMIN,
                 'name' => 'followup',
-                'rights' => READ | UPDATE | CREATE | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADDGROUPTICKET
-                    | ITILFollowup::ADDALLTICKET | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMY | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADD_AS_GROUP
+                    | ITILFollowup::ADDALLITEM | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER | ITILFollowup::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_ADMIN,
                 'name' => 'task',
-                'rights' => CommonITILTask::SEEPUBLIC | PURGE | CommonITILTask::UPDATEALL
-                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE,
+                'rights' => CommonITILTask::SEEPUBLIC | CommonITILTask::UPDATEMY | CommonITILTask::ADDMY | PURGE | CommonITILTask::UPDATEALL | CommonITILTask::ADD_AS_GROUP
+                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE | CommonITILTask::ADD_AS_OBSERVER | CommonITILTask::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_ADMIN,
                 'name' => 'projecttask',
@@ -6487,13 +6487,13 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_SUPER_ADMIN,
                 'name' => 'followup',
-                'rights' => READ | UPDATE | CREATE | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADDGROUPTICKET
-                    | ITILFollowup::ADDALLTICKET | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMY | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADD_AS_GROUP
+                    | ITILFollowup::ADDALLITEM | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER | ITILFollowup::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_SUPER_ADMIN,
                 'name' => 'task',
-                'rights' => CommonITILTask::SEEPUBLIC | PURGE | CommonITILTask::UPDATEALL
-                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE,
+                'rights' => CommonITILTask::SEEPUBLIC | CommonITILTask::UPDATEMY | CommonITILTask::ADDMY | PURGE | CommonITILTask::UPDATEALL | CommonITILTask::ADD_AS_GROUP
+                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE | CommonITILTask::ADD_AS_OBSERVER | CommonITILTask::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_SUPERVISOR,
                 'name' => 'project',
@@ -6785,12 +6785,13 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_HOTLINER,
                 'name' => 'followup',
-                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMYTICKET
-                    | ITILFollowup::ADDALLTICKET | ITILFollowup::SEEPRIVATE,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMY
+                    | ITILFollowup::ADDALLITEM | ITILFollowup::SEEPRIVATE,
             ], [
                 'profiles_id' => self::PROFILE_HOTLINER,
                 'name' => 'task',
-                'rights' => READ | CommonITILTask::SEEPRIVATE,
+                'rights' => CommonITILTask::SEEPUBLIC | CommonITILTask::UPDATEMY | CommonITILTask::ADDMY
+                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE,
             ], [
                 'profiles_id' => self::PROFILE_SUPER_ADMIN,
                 'name' => 'project',
@@ -7071,13 +7072,13 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_TECHNICIAN,
                 'name' => 'followup',
-                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMYTICKET
-                    | ITILFollowup::UPDATEALL | ITILFollowup::ADDALLTICKET | ITILFollowup::SEEPRIVATE,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMY | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADD_AS_GROUP
+                    | ITILFollowup::ADDALLITEM | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER | ITILFollowup::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_TECHNICIAN,
                 'name' => 'task',
-                'rights' => CommonITILTask::SEEPUBLIC | PURGE | CommonITILTask::UPDATEALL
-                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE,
+                'rights' => CommonITILTask::SEEPUBLIC | CommonITILTask::UPDATEMY | CommonITILTask::ADDMY | PURGE | CommonITILTask::UPDATEALL | CommonITILTask::ADD_AS_GROUP
+                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE | CommonITILTask::ADD_AS_OBSERVER | CommonITILTask::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_SELF_SERVICE,
                 'name' => 'project',
@@ -7362,13 +7363,13 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_SUPERVISOR,
                 'name' => 'followup',
-                'rights' => READ | UPDATE | CREATE | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADDGROUPTICKET
-                    | ITILFollowup::ADDALLTICKET | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER,
+                'rights' => ITILFollowup::SEEPUBLIC | ITILFollowup::UPDATEMY | ITILFollowup::ADDMY | PURGE | ITILFollowup::UPDATEALL | ITILFollowup::ADD_AS_GROUP
+                    | ITILFollowup::ADDALLITEM | ITILFollowup::SEEPRIVATE | ITILFollowup::ADD_AS_OBSERVER | ITILFollowup::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_SUPERVISOR,
                 'name' => 'task',
-                'rights' => CommonITILTask::SEEPUBLIC | PURGE | CommonITILTask::UPDATEALL
-                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE,
+                'rights' => CommonITILTask::SEEPUBLIC | CommonITILTask::UPDATEMY | CommonITILTask::ADDMY | PURGE | CommonITILTask::UPDATEALL | CommonITILTask::ADD_AS_GROUP
+                    | CommonITILTask::ADDALLITEM | CommonITILTask::SEEPRIVATE | CommonITILTask::ADD_AS_OBSERVER | CommonITILTask::ADD_AS_TECHNICIAN,
             ], [
                 'profiles_id' => self::PROFILE_SUPERVISOR,
                 'name' => 'queuednotification',