+++ install/migrations/update_10.0.2_to_10.0.3/inventory.php
@@ -40,4 +40,4 @@ use Glpi\Inventory\Conf;
  */
 
 //new right value for inventory
-$migration->updateRight('inventory', READ | Conf::IMPORTFROMFILE | Conf::UPDATECONFIG, ['config' => UPDATE, 'inventory' => READ]);
+$migration->replaceRight('inventory', READ | Conf::IMPORTFROMFILE | Conf::UPDATECONFIG, ['config' => UPDATE, 'inventory' => READ]);