+++ install/migrations/update_10.0.x_to_11.0.0/itilfollowup.php
@@ -34,16 +34,15 @@
  */
 
 /**
- * @var \DBmysql $DB
  * @var \Migration $migration
  */
 
-$table = 'glpi_cables';
-if (!$DB->fieldExists($table, 'is_deleted', false)) {
-    $migration->addField($table, 'is_deleted', 'bool');
-    $migration->addKey($table, 'is_deleted');
-}
-
-$migration->replaceRight('cable_management', READ | UPDATE | CREATE | DELETE | PURGE, [
-    'cable_management' => READ | UPDATE | CREATE | PURGE
-]);
+// new right values for task
+$migration->giveRight(
+    'followup',
+    // ITILFollowup::ADD_AS_TECHNICIAN,
+    32768,
+    [
+        'ticket' => 32768, // Ticket::OWN
+    ]
+);