+++ src/CommonITILObject.php
@@ -976,32 +976,73 @@ abstract class CommonITILObject extends CommonDBTM
     public function canAddFollowups()
     {
         return (
-         (
-            Session::haveRight("followup", ITILFollowup::ADDMYTICKET)
-            && (
-               $this->isUser(CommonITILActor::REQUESTER, Session::getLoginUserID())
-               || (
-                  isset($this->fields["users_id_recipient"])
-                  && ($this->fields["users_id_recipient"] == Session::getLoginUserID())
-               )
+            (
+                Session::haveRight(ITILFollowup::$rightname, ITILFollowup::ADDMY)
+                && (
+                    $this->isUser(CommonITILActor::REQUESTER, Session::getLoginUserID())
+                    || (
+                        isset($this->fields["users_id_recipient"])
+                        && ($this->fields["users_id_recipient"] == Session::getLoginUserID())
+                    )
+                )
             )
-         )
-         || (
-            Session::haveRight("followup", ITILFollowup::ADD_AS_OBSERVER)
-            && $this->isUser(CommonITILActor::OBSERVER, Session::getLoginUserID())
-         )
-         || Session::haveRight('followup', ITILFollowup::ADDALLTICKET)
-         || (
-            Session::haveRight('followup', ITILFollowup::ADDGROUPTICKET)
-            && isset($_SESSION["glpigroups"])
-            && $this->haveAGroup(CommonITILActor::REQUESTER, $_SESSION['glpigroups'])
-         )
-         || $this->isUser(CommonITILActor::ASSIGN, Session::getLoginUserID())
-         || (
-            isset($_SESSION["glpigroups"])
-            && $this->haveAGroup(CommonITILActor::ASSIGN, $_SESSION['glpigroups'])
-         )
-         || $this->isUserValidationRequested(Session::getLoginUserID(), true)
+            || (
+                Session::haveRight(ITILFollowup::$rightname, ITILFollowup::ADD_AS_OBSERVER)
+                && $this->isUser(CommonITILActor::OBSERVER, Session::getLoginUserID())
+            )
+            || Session::haveRight(ITILFollowup::$rightname, ITILFollowup::ADDALLITEM)
+            || (
+                Session::haveRight(ITILFollowup::$rightname, ITILFollowup::ADD_AS_GROUP)
+                && isset($_SESSION["glpigroups"])
+                && $this->haveAGroup(CommonITILActor::REQUESTER, $_SESSION['glpigroups'])
+            )
+            || (
+                Session::haveRight(ITILFollowup::$rightname, ITILFollowup::ADD_AS_TECHNICIAN)
+                && (
+                    $this->isUser(CommonITILActor::ASSIGN, Session::getLoginUserID())
+                    || (
+                        isset($_SESSION["glpigroups"])
+                        && $this->haveAGroup(CommonITILActor::ASSIGN, $_SESSION['glpigroups'])
+                    )
+                )
+            )
+            || $this->isUserValidationRequested(Session::getLoginUserID(), true)
+        );
+    }
+
+    public function canAddTasks()
+    {
+        return (
+            (
+                Session::haveRight(CommonITILTask::$rightname, CommonITILTask::ADDMY)
+                && (
+                    $this->isUser(CommonITILActor::REQUESTER, Session::getLoginUserID())
+                    || (
+                        isset($this->fields["users_id_recipient"])
+                        && ($this->fields["users_id_recipient"] == Session::getLoginUserID())
+                    )
+                )
+            )
+            || (
+                Session::haveRight(CommonITILTask::$rightname, CommonITILTask::ADD_AS_OBSERVER)
+                && $this->isUser(CommonITILActor::OBSERVER, Session::getLoginUserID())
+            )
+            || Session::haveRight(CommonITILTask::$rightname, CommonITILTask::ADDALLITEM)
+            || (
+                Session::haveRight(CommonITILTask::$rightname, CommonITILTask::ADD_AS_GROUP)
+                && isset($_SESSION["glpigroups"])
+                && $this->haveAGroup(CommonITILActor::REQUESTER, $_SESSION['glpigroups'])
+            )
+            || (
+                Session::haveRight(CommonITILTask::$rightname, CommonITILTask::ADD_AS_TECHNICIAN)
+                && (
+                    $this->isUser(CommonITILActor::ASSIGN, Session::getLoginUserID())
+                    || (
+                        isset($_SESSION["glpigroups"])
+                        && $this->haveAGroup(CommonITILActor::ASSIGN, $_SESSION['glpigroups'])
+                    )
+                )
+            )
         );
     }